Veryfies if CordovaSim proccess is started after Cordova project is started
