+++ tests/org.jboss.tools.aerogear.ui.bot.test/src/org/jboss/tools/aerogear/ui/bot/test/app/RunWithCordovaSim.java
@@ -3,6 +3,7 @@ package org.jboss.tools.aerogear.ui.bot.test.app;
 import java.util.Date;
 
 import org.jboss.tools.aerogear.ui.bot.test.AerogearBotTest;
+import org.jboss.tools.ui.bot.ext.SWTUtilExt;
 import org.jboss.tools.ui.bot.ext.config.Annotations.Require;
 import org.junit.After;
 import org.junit.Before;
@@ -22,8 +23,13 @@ public class RunWithCordovaSim extends AerogearBotTest {
 	@Test
 	public void canRunWithCordovaSim() {
 		projectExplorer.selectProject(CORDOVA_APP);
-
-		runTreeItemWithCordovaSim(bot.tree().expandNode(CORDOVA_APP));
+    final String cordovaSimProcessName = "CordovaSimRunner";
+    int countBrowserSimmProcesses = SWTUtilExt.countJavaProcess(cordovaSimProcessName);
+    // this also asserts that CordovaSim runs without error within JBT
+    runTreeItemWithCordovaSim(bot.tree().expandNode(CORDOVA_APP));
+    assertTrue("No new CordovaSimm process was started",countBrowserSimmProcesses + 1 == SWTUtilExt.countJavaProcess(cordovaSimProcessName));
+    // currently there is no way how to close CordovaSim within running JBT
+    // CordovaSim is automatically closed when JBT are
 	}
 
 	@After
